move all activities to ui package
