--- app/src/main/java/com/arctouch/codechallenge/activity/base/BaseView.kt
@@ -1,4 +1,0 @@
-package com.arctouch.codechallenge.activity.base
-
-interface BaseView {
-}
