--- app/src/main/java/com/arctouch/codechallenge/activity/details/DetailsPresenter.kt
@@ -1,6 +1,0 @@
-package com.arctouch.codechallenge.activity.details
-
-import com.arctouch.codechallenge.activity.base.BasePresenterContract
-
-interface DetailsPresenter: BasePresenterContract {
-}
