+++ app/src/main/java/com/arctouch/codechallenge/di/component/NetworkComponent.kt
@@ -1,6 +1,6 @@
 package com.arctouch.codechallenge.di.component
 
-import com.arctouch.codechallenge.activity.base.BaseNetworkActivity
+import com.arctouch.codechallenge.ui.activity.base.BaseNetworkActivity
 import com.arctouch.codechallenge.di.module.NetworkModule
 import dagger.Component
