+++ app/src/main/java/com/arctouch/codechallenge/ui/activity/base/BasePresenterImpl.kt
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.activity.base
+package com.arctouch.codechallenge.ui.activity.base
 
 import android.os.Bundle
