+++ app/src/main/java/com/arctouch/codechallenge/ui/activity/base/BaseView.kt
@@ -1,0 +1,4 @@
+package com.arctouch.codechallenge.ui.activity.base
+
+interface BaseView {
+}
