+++ app/src/main/java/com/arctouch/codechallenge/ui/activity/base/LoaderView.kt
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.activity.base
+package com.arctouch.codechallenge.ui.activity.base
 
 interface LoaderView {
