+++ app/src/main/java/com/arctouch/codechallenge/ui/activity/details/DetailsPresenter.kt
@@ -1,0 +1,6 @@
+package com.arctouch.codechallenge.ui.activity.details
+
+import com.arctouch.codechallenge.ui.activity.base.BasePresenterContract
+
+interface DetailsPresenter: BasePresenterContract {
+}
